Make return value of nside_to_level a 64-bit int
It is very common to use the return value as a bit-shift argument
which would cause truncation if the user failed to first convert
the return value to a 64-bit integer.

See https://git.ligo.org/emfollow/userguide/-/merge_requests/332.

+++ astropy_healpix/core.py
@@ -130,7 +130,7 @@ def uniq_to_level_ipix(uniq):
     good = uniq >= 4
     i = _core.bit_scan_reverse(uniq) >> 1
     level = i - 1
-    ipix = uniq - (np.int64(1) << (i << 1))
+    ipix = uniq - (1 << (i << 1))
     return np.where(good, level, -1), np.where(good, ipix, -1)
